requirements.txt
llama-cpp-python
pypdf
python-dotenv
